requirements.txt
streamlit==1.31.0
langchain==0.1.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0